+++ dse_do_utils/optimizationengine.py
@@ -32,14 +32,18 @@ except ImportError:
     from scenariomanager import ScenarioManager
     from datamanager import DataManager
 
+from typing import TypeVar, Generic
 
-class OptimizationEngine(object):
-    def __init__(self, data_manager: Optional[DataManager] = None, name: str = "MyOptimizationEngine",
+DM = TypeVar('DM', bound='DataManager')
+
+
+class OptimizationEngine(Generic[DM]):
+    def __init__(self, data_manager: Optional[DM] = None, name: str = "MyOptimizationEngine",
                  solve_kwargs = None, export_lp: bool = False, export_sav: bool = False, export_lp_path: str = None, is_cpo_model: bool = False):
         self.is_cpo_model = is_cpo_model
         # self.mdl: Model = Model(name=name)
         self.mdl: Union[Model, cp.CpoModel] = self.create_do_model(name=name, is_cpo_model=is_cpo_model)
-        self.dm = data_manager
+        self.dm: DM = data_manager
         self.solve_kwargs = solve_kwargs  # TODO: use in this.solve()
         self.export_lp = export_lp
         self.export_sav = export_sav  # TODO: add export to sav
@@ -89,17 +93,17 @@ class OptimizationEngine(object):
     @staticmethod
     def integer_var_series_s(mdl: docplex.mp.model, df: pd.DataFrame, **kargs) -> pd.Series:
         """Returns pd.Series[IntegerVarType]"""
-        return pd.Series(mdl.integer_var_list(df.index, **kargs), index=df.index)
+        return pd.Series(mdl.integer_var_list(df.index, **kargs), index=df.index, dtype='object')
 
     @staticmethod
     def continuous_var_series_s(mdl: docplex.mp.model, df: pd.DataFrame, **kargs) -> pd.Series:
         """Returns pd.Series[ContinuousVarType]."""
-        return pd.Series(mdl.continuous_var_list(df.index, **kargs), index=df.index)
+        return pd.Series(mdl.continuous_var_list(df.index, **kargs), index=df.index, dtype='object')
 
     @staticmethod
     def binary_var_series_s(mdl: docplex.mp.model, df: pd.DataFrame, **kargs) -> pd.Series:
         """Returns pd.Series[BinaryVarType]"""
-        return pd.Series(mdl.binary_var_list(df.index, **kargs), index=df.index)
+        return pd.Series(mdl.binary_var_list(df.index, **kargs), index=df.index, dtype='object')
 
     def semicontinuous_var_series(self, df, lb, **kargs) -> pd.Series:
         """Returns pd.Series[SemiContinuousVarType]"""
@@ -108,7 +112,7 @@ class OptimizationEngine(object):
     @staticmethod
     def semicontinuous_var_series_s(mdl: docplex.mp.model, df: pd.DataFrame, lb, **kargs) -> pd.Series:
         """Returns pd.Series[SemiContinuousVarType]."""
-        return pd.Series(mdl.semicontinuous_var_list(df.index, lb, **kargs), index=df.index)
+        return pd.Series(mdl.semicontinuous_var_list(df.index, lb, **kargs), index=df.index, dtype='object')
 
     def semiinteger_var_series(self, df, lb, **kargs) -> pd.Series:
         """Returns pd.Series[SemiIntegerVarType]"""
@@ -117,7 +121,7 @@ class OptimizationEngine(object):
     @staticmethod
     def semiinteger_var_series_s(mdl: docplex.mp.model, df: pd.DataFrame, lb, **kargs) -> pd.Series:
         """Returns pd.Series[SemiIntegerVarType]."""
-        return pd.Series(mdl.semiinteger_var_list(df.index, lb, **kargs), index=df.index)
+        return pd.Series(mdl.semiinteger_var_list(df.index, lb, **kargs), index=df.index, dtype='object')
 
     def solve(self, refine_conflict: bool = False, **kwargs) -> docplex.mp.solution.SolveSolution:
         # TODO: enable export_as_lp_path()?