+++ dse_do_utils/scenariorunner.py
@@ -5,12 +5,17 @@ from copy import deepcopy
 from dataclasses import dataclass
 
 import pandas as pd
+from dse_do_utils.core.core01_optimization_engine import Core01OptimizationEngine
+
+from dse_do_utils.core.core01_data_manager import Core01DataManager
 
 from dse_do_utils import ScenarioManager, OptimizationEngine
 from dse_do_utils.datamanager import Inputs, Outputs, DataManager
 from dse_do_utils.scenariodbmanager import ScenarioDbManager
 from logging import Logger, getLogger
-from typing import Any, Dict, Optional, Tuple, NamedTuple, Type, List
+from typing import Any, Dict, Optional, Tuple, NamedTuple, Type, List, Union
+
+from dse_do_utils.scenariomanager import Platform
 
 
 @dataclass  # (frozen=True)
@@ -31,6 +36,7 @@ class RunConfig:
     log_level: str = 'DEBUG'  # 'DEBUG'
     export_lp: bool = False
     export_sav: bool = False
+    enable_refine_conflict: bool = False
     export_lp_path: str = ''
     do_model_name: str = None
     template_scenario_name: Optional[str] = None  # 'TemplateScenario'
@@ -55,7 +61,7 @@ class ScenarioGenerator():
     def __init__(self,
                  inputs: Inputs,
                  scenario_config: ScenarioConfig) -> None:
-        self._logger: Logger = getLogger(self.__class__.__name__)
+        self._logger: Logger = getLogger(__name__)
         self.inputs: Inputs = inputs.copy()  # Only copy of dict
         self.scenario_config: ScenarioConfig = scenario_config
 
@@ -79,9 +85,19 @@ class ScenarioGenerator():
         """Applies overrides to the Parameter table based on the ScenarioConfig.parameters.
         """
         if self.scenario_config.parameters is None:
-            df = self.inputs['Parameter']
+            if 'Parameter' in self.inputs.keys():
+                df = self.inputs['Parameter']
+            elif 'Parameters' in self.inputs.keys():
+                df = self.inputs['Parameters']
+            else:
+                df = pd.DataFrame(columns=['param', 'value']).set_index('param')
         else:
-            df = self.inputs['Parameter'].copy().set_index(['param'])
+            if 'Parameter' in self.inputs.keys():
+                df = self.inputs['Parameter'].copy().set_index(['param'])
+            elif 'Parameters' in self.inputs.keys():
+                df = self.inputs['Parameters'].copy().set_index(['param'])
+            else:
+                df = pd.DataFrame(columns=['param', 'value']).set_index('param')
             for param, value in self.scenario_config.parameters.items():
                 df.at[param, 'value'] = value
         return df
@@ -93,28 +109,29 @@ class ScenarioRunner:
     """
     def __init__(self,
                  scenario_db_manager: ScenarioDbManager,
-                 optimization_engine_class: Type[OptimizationEngine],
-                 data_manager_class: Type[DataManager],
+                 optimization_engine_class: Type[Core01OptimizationEngine],
+                 data_manager_class: Type[Core01DataManager],
                  scenario_db_manager_class: Type[ScenarioDbManager],  # For the SQLite data check
                  scenario_generator_class: Optional[Type[ScenarioGenerator]] = None,
                  do_model_name: str = 'my_model',
                  schema: Optional[str] = None,
                  # use_scenario_db: bool = True,
                  local_root: Optional[str] = None,
-                 local_platform: Optional[int] = None,
+                 local_platform: Optional[Union[int, Platform]] = None,
                  data_directory: Optional[str] = None) -> None:
 
         self.scenario_db_manager: ScenarioDbManager = scenario_db_manager
-        self.optimization_engine_class: Type[OptimizationEngine] = optimization_engine_class
-        self.data_manager_class = data_manager_class
+        self.optimization_engine_class: Type[Core01OptimizationEngine] = optimization_engine_class
+        self.data_manager_class: Type[Core01DataManager] = data_manager_class
         self.scenario_db_manager_class = scenario_db_manager_class
         self.scenario_generator_class = scenario_generator_class
 
-        self.optimization_engine: OptimizationEngine = None  # To be set in run.
-        self.data_manager: DataManager = None  # To be set in run.
-        self.sqlite_scenario_db_manager: ScenarioDbManager = None  # To be set in run.
+        self.optimization_engine: Optional[OptimizationEngine] = None  # To be set in run.
+        self.data_manager: Optional[DataManager] = None  # To be set in run.
+        self.sqlite_scenario_db_manager: Optional[ScenarioDbManager] = None  # To be set in run.
 
-        self._logger: Logger = getLogger(self.__class__.__name__)
+        # self._logger: Logger = getLogger(self.__class__.__name__)
+        self._logger: Logger = getLogger(__name__)
         self.schema: Optional[str] = schema
         self.do_model_name: str = do_model_name
         # self.use_scenario_db: bool = use_scenario_db  # TODO: VT20220906: remove, doesn't seem to be used?
@@ -171,7 +188,7 @@ class ScenarioRunner:
         #         'Either base_inputs or excel_file_name should be provided.')
 
         # Generate scenario
-        self._logger.info(f'Generating scenario {scenario_name}')
+        self._logger.debug(f'Generating scenario {scenario_name}')
         inputs = self.generate_scenario(base_inputs, scenario_config)
 
         # Data check
@@ -199,7 +216,7 @@ class ScenarioRunner:
         if run_config.write_output_to_excel:
             self.write_output_data_to_excel(inputs, outputs, scenario_name)
 
-        self._logger.info(f'Done with {scenario_config.scenario_name}')
+        self._logger.debug(f'Done with {scenario_config.scenario_name}')
 
         return outputs
 
@@ -210,7 +227,7 @@ class ScenarioRunner:
     def _load_base_inputs(self, excel_file_name, base_inputs):
         # Load base inputs
         if excel_file_name and not base_inputs:
-            self._logger.info('Loading data from the excel file')
+            self._logger.info(f'Loading data from the excel file {excel_file_name}')
             inputs = self.load_input_data_from_excel(excel_file_name)
         elif not excel_file_name and base_inputs:
             inputs = base_inputs
@@ -390,13 +407,14 @@ class ScenarioRunner:
             export_lp=run_config.export_lp,
             export_sav=run_config.export_sav,
             export_lp_path=run_config.export_lp_path,
+            enable_refine_conflict=run_config.enable_refine_conflict
         )
 
         return self.optimization_engine.run()
 
     def insert_outputs_in_db(self, inputs: Inputs, outputs: Outputs, run_config: RunConfig, scenario_name: str):
         self._logger.info('Inserting outputs into the database')
-        if run_config.insert_inputs_in_db:
+        if run_config.insert_outputs_in_db:
             self.scenario_db_manager.update_scenario_output_tables_in_db(scenario_name, outputs)
         else:
             self.scenario_db_manager.replace_scenario_in_db(scenario_name, inputs, outputs)